Client credential provisioning: User agent mediates CSR from the client (#76)
In case the plgd d2d client or other application using device/client Go library cannot access the plgd hub API directly (e.g. it’s running in a different network as a web service), but the user’s PC can, the CSR can be propagated through the browser to the plgd hub API and the response from the certificate authority service can be returned back over the browser to the client. This flow delivers the Identity Certificate to the client, without private key leaving the client. The client can communicate securely with any devices which is part of the same security domain right after receiving the Identity certificate and its initialization.

Co-authored-by: Patrik Matiaško <patrik.matiasko@gmail.com>