Support Gateway with Multiple Devices using the Same Endpoint
Allows multiple devices to be connected to the same endpoint.
By utilizing a single endpoint, multiple devices can now be exposed and controlled efficiently.
To achieve this, a new query parameter, "di"(DeviceID) has been added to each command.
This parameter is used to specify the target device for the command execution.